client_id is already mandated by OAuth
Originally I wanted to just wrap the "it must still be sent and valid"
in parens (since the resulting sentence was extremely weird and read
like rambling).

However, client_id is already REQUIRED by the OAuth RFC, IMO it's not
necessary to reiterate that requirement here.

This reapplies #120 (8e36571) after it was accidentally reverted in
9e14a6e87a1209f85958421cb984348e5c5731e0